Fixed not being able to have multiple of the same card in a set

+++ Gwent.NET/Commands/Command.cs
@@ -13,7 +13,7 @@ namespace Gwent.NET.Commands
             Events = new List<Event>();
         }
 
-        public int SenderUserId { get; set; }
+        public long SenderUserId { get; set; }
         public ICollection<Event> Events { get; set; }
         public State NextState { get; set; }
