+++ Gwent.NET/DTOs/PlayerDto.cs
@@ -1,12 +1,12 @@
 ﻿using System.Collections.Generic;
-using Gwent.NET.Model;
+using System.Runtime.CompilerServices;
 using Gwent.NET.Model.Enums;
 
 namespace Gwent.NET.DTOs
 {
     public class PlayerDto
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Name { get; set; }
         public bool IsLobbyOwner { get; set; }
 
@@ -15,20 +15,20 @@ namespace Gwent.NET.DTOs
         public int Lives { get; set; }
 
         public GwintFaction Faction { get; set; }
-        public int? BattleKingCard { get; set; }
+        public long? BattleKingCard { get; set; }
         public bool CanUseBattleKingCard { get; set; }
         
         public int HandCardCount { get; set; }
         public int DeckCardCount { get; set; }
 
-        public List<int> HandCards { get; set; }
-        public List<int> GraveyardCards { get; set; }
-        public List<int> MeleeCards { get; set; }
-        public List<int> RangeCards { get; set; }
-        public List<int> SiegeCards { get; set; }
-        public List<int> WeatherCards { get; set; }
-        public List<int> MeleeModifiers { get; set; }
-        public List<int> RangedModifiers { get; set; }
-        public List<int> SiegeModifiers { get; set; }
+        public List<long> HandCards { get; set; }
+        public List<long> GraveyardCards { get; set; }
+        public List<long> MeleeCards { get; set; }
+        public List<long> RangeCards { get; set; }
+        public List<long> SiegeCards { get; set; }
+        public List<long> WeatherCards { get; set; }
+        public List<long> MeleeModifiers { get; set; }
+        public List<long> RangedModifiers { get; set; }
+        public List<long> SiegeModifiers { get; set; }
     }
 }
